flashcards-lang/requirements.txt
flask==2.0.1
Pillow==8.4.0
requests==2.31.0
deep-translator==1.9.1
werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
